Added curses in example.py, fixed wrong datatypes for the gyroscope

+++ embedded/mega2560/mpu_dev.h
@@ -16,7 +16,7 @@ public:
 
 	float temp, pressure, altitude;
 
-	float ax, ay, az, gx, gy, gz, mx, my, mz;
+	int16_t ax, ay, az, gx, gy, gz, mx, my, mz;
 	
 	void update();
 	int send();
